+++ src/test/java/se/bjurr/prnfb/admin/PrnfbPullRequestEventListenerTest.java
@@ -4,11 +4,15 @@ import static com.atlassian.bitbucket.pull.PullRequestAction.COMMENTED;
 import static com.atlassian.bitbucket.pull.PullRequestAction.MERGED;
 import static com.atlassian.bitbucket.pull.PullRequestAction.OPENED;
 import static com.atlassian.bitbucket.pull.PullRequestAction.RESCOPED;
+import static com.atlassian.bitbucket.pull.PullRequestRole.AUTHOR;
+import static com.atlassian.bitbucket.pull.PullRequestRole.PARTICIPANT;
 import static com.atlassian.bitbucket.pull.PullRequestState.DECLINED;
 import static com.google.common.base.Charsets.UTF_8;
 import static com.google.common.base.Joiner.on;
 import static com.google.common.collect.Lists.newArrayList;
 import static com.google.common.io.Resources.getResource;
+import static java.lang.Boolean.FALSE;
+import static java.lang.Boolean.TRUE;
 import static java.util.Collections.sort;
 import static javax.ws.rs.core.HttpHeaders.AUTHORIZATION;
 import static org.junit.Assert.assertEquals;
@@ -107,30 +111,35 @@ public class PrnfbPullRequestEventListenerTest {
       notificationBuilder() //
         .withFieldValue(
           url,
-          "http://bjurr.se/?PULL_REQUEST_FROM_HASH=${PULL_REQUEST_FROM_HASH}&PULL_REQUEST_TO_HASH=${PULL_REQUEST_TO_HASH}&PULL_REQUEST_FROM_REPO_SLUG=${PULL_REQUEST_FROM_REPO_SLUG}&PULL_REQUEST_TO_REPO_SLUG=${PULL_REQUEST_TO_REPO_SLUG}") //
+          "http://bjurr.se/?PULL_REQUEST_FROM_HASH=${PULL_REQUEST_FROM_HASH}&PULL_REQUEST_TO_HASH=${PULL_REQUEST_TO_HASH}&PULL_REQUEST_FROM_REPO_SLUG=${PULL_REQUEST_FROM_REPO_SLUG}&PULL_REQUEST_TO_REPO_SLUG=${PULL_REQUEST_TO_REPO_SLUG}&revapp=${PULL_REQUEST_REVIEWERS_APPROVED_COUNT}&partapp=${PULL_REQUEST_PARTICIPANTS_APPROVED_COUNT}") //
         .withFieldValue(events, OPENED.name()) //
         .build() //
     ) //
     .store() //
     .trigger( //
-      pullRequestEventBuilder() //
-        .withFromRef( //
-          pullRequestRefBuilder() //
-            .withHash("cde456") //
-            .withRepositorySlug("fromslug") //
-        ) //
-        .withToRef( //
+      pullRequestEventBuilder().withFromRef( //
+        pullRequestRefBuilder() //
+          .withHash("cde456") //
+          .withRepositorySlug("fromslug") //
+        ).withToRef( //
           pullRequestRefBuilder() //
             .withHash("asd123") //
             .withRepositorySlug("toslug") //
         ) //
         .withPullRequestId(10L) //
         .withPullRequestAction(OPENED) //
+        .withParticipant(PARTICIPANT, TRUE) //
+        .withParticipant(PARTICIPANT, TRUE) //
+        .withParticipant(PARTICIPANT, TRUE) //
+        .withParticipant(PARTICIPANT, FALSE) //
+        .withParticipant(AUTHOR, TRUE) //
+        .withParticipant(AUTHOR, TRUE) //
+        .withParticipant(AUTHOR, FALSE) //
         .build() //
     ) //
     .invokedUrl(
       0,
-      "http://bjurr.se/?PULL_REQUEST_FROM_HASH=cde456&PULL_REQUEST_TO_HASH=asd123&PULL_REQUEST_FROM_REPO_SLUG=fromslug&PULL_REQUEST_TO_REPO_SLUG=toslug") //
+      "http://bjurr.se/?PULL_REQUEST_FROM_HASH=cde456&PULL_REQUEST_TO_HASH=asd123&PULL_REQUEST_FROM_REPO_SLUG=fromslug&PULL_REQUEST_TO_REPO_SLUG=toslug&revapp=2&partapp=3") //
     .invokedMethod(GET);
  }